requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Login==0.6.3
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
